# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: KEMRI-CGHR Influenza Program Clinical Data Management System
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /screening:
    get:
      operationId: listScreenings
      tags: [screening]
      summary: List all screenings
      parameters:
        - name: facility
          in: query
          schema:
            type: string
        - name: eligible
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of screenings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Screening"
    post:
      operationId: createScreening
      tags: [screening]
      summary: Create a screening record
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ScreeningInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Screening"
  /screening/{screeningId}:
    get:
      operationId: getScreening
      tags: [screening]
      summary: Get a screening record
      parameters:
        - name: screeningId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Screening record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Screening"
        "404":
          description: Not found
    patch:
      operationId: updateScreening
      tags: [screening]
      summary: Update a screening record
      parameters:
        - name: screeningId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ScreeningUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Screening"
    delete:
      operationId: deleteScreening
      tags: [screening]
      summary: Delete a screening record
      parameters:
        - name: screeningId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DeleteInput"
      responses:
        "200":
          description: Deleted
...
  /enrolment:
    get:
      operationId: listEnrolments
      tags: [enrolment]
      summary: List all enrolments
      parameters:
        - name: facility
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of enrolments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Enrolment"
    post:
      operationId: createEnrolment
      tags: [enrolment]
      summary: Create an enrolment record
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EnrolmentInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Enrolment"
  /enrolment/{screeningId}:
    get:
      operationId: getEnrolment
      tags: [enrolment]
      summary: Get an enrolment record by screening ID
      parameters:
        - name: screeningId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Enrolment record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Enrolment"
        "404":
          description: Not found
    patch:
      operationId: updateEnrolment
      tags: [enrolment]
      summary: Update an enrolment record
      parameters:
        - name: screeningId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EnrolmentUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Enrolment"
    delete:
      operationId: deleteEnrolment
      tags: [enrolment]
      summary: Delete an enrolment record
      parameters:
        - name: screeningId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DeleteInput"
      responses:
        "200":
          description: Deleted
...
components:
  schemas:
...
    DeleteInput:
      type: object
      required: [reason]
      properties:
        reason:
          type: string

    # Screening
    Screening:
      type: object
      required: [screeningId, interviewDate, healthFacility, createdBy, createdAt]
      properties:
        screeningId:
          type: string
        interviewDate:
          type: string
        healthFacility:
          type: string
        dob:
          type: ["string", "null"]
        ageYears:
          type: ["integer", "null"]
        ageMonths:
          type: ["integer", "null"]
        height:
          type: ["number", "null"]
        weight:
          type: ["number", "null"]
        temperature:
          type: ["number", "null"]
        respRate:
          type: ["integer", "null"]
        pulseRate:
          type: ["integer", "null"]
        bloodPressure:
          type: ["string", "null"]
        lmp:
          type: ["string", "null"]
        fundalHeight:
          type: ["number", "null"]
        inclusion1:
          type: ["string", "null"]
        inclusion2:
          type: ["string", "null"]
        inclusion3:
          type: ["string", "null"]
        inclusion4:
          type: ["string", "null"]
        inclusion5:
          type: ["string", "null"]
        exclusion1:
          type: ["string", "null"]
        exclusion2:
          type: ["string", "null"]
        exclusion3:
          type: ["string", "null"]
        eligible:
          type: ["string", "null"]
        consented:
          type: ["string", "null"]
        bmi:
          type: ["number", "null"]
        createdBy:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: ["string", "null"]
    ScreeningInput:
      type: object
      required: [interviewDate, healthFacility]
      properties:
        interviewDate:
          type: string
        healthFacility:
          type: string
        dob:
          type: string
        ageYears:
          type: integer
        ageMonths:
          type: integer
        height:
          type: number
        weight:
          type: number
        temperature:
          type: number
        respRate:
          type: integer
        pulseRate:
          type: integer
        bloodPressure:
          type: string
        lmp:
          type: string
        fundalHeight:
          type: number
        inclusion1:
          type: string
        inclusion2:
          type: string
        inclusion3:
          type: string
        inclusion4:
          type: string
        inclusion5:
          type: string
        exclusion1:
          type: string
        exclusion2:
          type: string
        exclusion3:
          type: string
        eligible:
          type: string
        consented:
          type: string
    ScreeningUpdate:
      type: object
      properties:
        interviewDate:
          type: string
        healthFacility:
          type: string
        dob:
          type: string
        ageYears:
          type: integer
        ageMonths:
          type: integer
        height:
          type: number
        weight:
          type: number
        temperature:
          type: number
        respRate:
          type: integer
        pulseRate:
          type: integer
        bloodPressure:
          type: string
        lmp:
          type: string
        fundalHeight:
          type: number
        inclusion1:
          type: string
        inclusion2:
          type: string
        inclusion3:
          type: string
        inclusion4:
          type: string
        inclusion5:
          type: string
        exclusion1:
          type: string
        exclusion2:
          type: string
        exclusion3:
          type: string
        eligible:
          type: string
        consented:
          type: string
        reason:
          type: string
...
    Enrolment:
      type: object
      required: [id, screeningId, createdBy, createdAt]
      properties:
        id:
          type: integer
        screeningId:
          type: string
        maritalStatus:
          type: ["string", "null"]
        husbandName:
          type: ["string", "null"]
        village:
          type: ["string", "null"]
        education:
          type: ["string", "null"]
        occupation:
          type: ["string", "null"]
        height:
          type: ["number", "null"]
        weight:
          type: ["number", "null"]
        temperature:
          type: ["number", "null"]
        respRate:
          type: ["integer", "null"]
        pulseRate:
          type: ["integer", "null"]
        bloodPressure:
          type: ["string", "null"]
        gestationalAge:
          type: ["integer", "null"]
        bmi:
          type: ["number", "null"]
        createdBy:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: ["string", "null"]
    EnrolmentInput:
      type: object
      required: [screeningId]
      properties:
        screeningId:
          type: string
        maritalStatus:
          type: string
        husbandName:
          type: string
        village:
          type: string
        education:
          type: string
        occupation:
          type: string
        height:
          type: number
        weight:
          type: number
        temperature:
          type: number
        respRate:
          type: integer
        pulseRate:
          type: integer
        bloodPressure:
          type: string
        gestationalAge:
          type: integer
    EnrolmentUpdate:
      type: object
      properties:
        maritalStatus:
          type: string
        husbandName:
          type: string
        village:
          type: string
        education:
          type: string
        occupation:
          type: string
        height:
          type: number
        weight:
          type: number
        temperature:
          type: number
        respRate:
          type: integer
        pulseRate:
          type: integer
        bloodPressure:
          type: string
        gestationalAge:
          type: integer
        reason:
          type: string

    # ANC Visit
...